# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/comment/create": {
            "post": {
                "description": "Create comment from payload with POST method; need Authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Create comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpmodel.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comment/delete/{comment_id}": {
            "delete": {
                "description": "Delete comment by comment id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Delete Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the comment to delete",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comment/get/{comment_id}": {
            "get": {
                "description": "Get comment by id. This id must be taken from path and need authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Get Comment BY ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the comment to get comment",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comment/list/{post_id}": {
            "get": {
                "description": "List comments with giving post_id which is requested from params. Authenticates given user by giving an access jwttoken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "List comments by post id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the post to list comments",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comment/update": {
            "put": {
                "description": "Update comment from payload with PUT method; need Authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Update Comment By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/comment/update/{comment_id}/image": {
            "put": {
                "description": "Update comment image with authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth token of logged-in user.",
                        "name": "X-Auth-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the comment to update the image for",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "The image file to upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpmodel.UpdateImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comment.CreateRequest": {
            "type": "object",
            "required": [
                "body",
                "post_id"
            ],
            "properties": {
                "post_id": {
                    "description": "ID of the post",
                    "type": "integer",
                    "x-order": "1",
                    "example": 1
                },
                "body": {
                    "description": "Body of the post",
                    "type": "string",
                    "x-order": "2",
                    "example": "New Post..."
                },
                "parent_id": {
                    "description": "ID of the parent comment id",
                    "type": "integer",
                    "x-order": "3",
                    "example": 1
                }
            }
        },
        "comment.UpdateRequest": {
            "type": "object",
            "required": [
                "body",
                "id",
                "post_id"
            ],
            "properties": {
                "id": {
                    "description": "ID of the comment",
                    "type": "integer",
                    "x-order": "1",
                    "example": 1
                },
                "post_id": {
                    "description": "ID of the post",
                    "type": "integer",
                    "x-order": "2",
                    "example": 3
                },
                "body": {
                    "description": "Body of the post",
                    "type": "string",
                    "x-order": "3",
                    "example": "3"
                }
            }
        },
        "entity.Comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "parenId": {
                    "type": "integer"
                },
                "post": {
                    "$ref": "#/definitions/entity.Post"
                },
                "postID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "entity.Post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profilePhoto": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "httpmodel.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the created model`",
                    "type": "integer",
                    "x-order": "1",
                    "example": 1
                }
            }
        },
        "httpmodel.UpdateImageResponse": {
            "type": "object",
            "properties": {
                "uploaded_file_name": {
                    "description": "Uploaded file name",
                    "type": "string",
                    "x-order": "1",
                    "example": "https://res-cdn.com/postId"
                }
            }
        },
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "x-order": "1",
                    "example": true
                },
                "status_code": {
                    "type": "integer",
                    "x-order": "2",
                    "example": 200
                },
                "error": {
                    "x-order": "4",
                    "$ref": "#/definitions/httpresponse.ResponseError"
                },
                "data": {
                    "x-order": "5"
                }
            }
        },
        "httpresponse.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1",
                    "example": "NOT_FOUND"
                },
                "detail": {
                    "type": "string",
                    "x-order": "2",
                    "example": "user not found"
                }
            }
        },
...
    }
}